move handleFileRelatedDate to service layer
