+++ src/main/java/ca/gc/aafc/objectstore/api/respository/ObjectStoreResourceRepository.java
@@ -38,8 +38,7 @@ import java.util.UUID;
 @Repository
 @Transactional
 public class ObjectStoreResourceRepository
-  extends DinaRepository<ObjectStoreMetadataDto, ObjectStoreMetadata>
-  implements ObjectStoreMetadataReadService {
+  extends DinaRepository<ObjectStoreMetadataDto, ObjectStoreMetadata> {
 
   private final ObjectStoreMetaDataService dinaService;
   private final DinaAuthenticatedUser authenticatedUser;
@@ -77,10 +76,7 @@ public class ObjectStoreResourceRepository
   @Override
   @SuppressWarnings("unchecked")
   public <S extends ObjectStoreMetadataDto> S save(S resource) {
-    handleFileRelatedData(resource);
-    loadObjectStoreMetadata(resource.getUuid()).ifPresent(objectStoreMetadata ->
-      resource.setManagedAttributeMap(
-        MetadataToManagedAttributeMapRepository.getAttributeMapFromMetadata(objectStoreMetadata)));
+
     S dto = super.save(resource);
     return (S) this.findOne(dto.getUuid(), new QuerySpec(ObjectStoreMetadataDto.class));
   }
@@ -119,25 +115,9 @@ public class ObjectStoreResourceRepository
     return super.findAll(jpaFriendlyQuerySpec);
   }
 
-  @Override
-  public Optional<ObjectStoreMetadata> loadObjectStoreMetadata(UUID id) {
-    return Optional.ofNullable(dinaService.findOne(id, ObjectStoreMetadata.class));
-  }
-
-  @Override
-  public Optional<ObjectStoreMetadata> loadObjectStoreMetadataByFileId(UUID fileId) {
-    return dinaService.findAll(
-      ObjectStoreMetadata.class,
-      (cb, root) -> new Predicate[]{cb.equal(root.get("fileIdentifier"), fileId)}
-      , null, 0, 1)
-      .stream().findFirst();
-  }
-
   @SuppressWarnings("unchecked")
   @Override
   public ObjectStoreMetadataDto create(ObjectStoreMetadataDto resource) {
-
-    handleFileRelatedData(resource);
 
     resource.setCreatedBy(authenticatedUser.getUsername());
     ObjectStoreMetadataDto created = super.create(resource);
@@ -146,38 +126,6 @@ public class ObjectStoreResourceRepository
       created.getUuid(),
       new QuerySpec(ObjectStoreMetadataDto.class)
     );
-  }
-
-  /**
-   * Method responsible for dealing with validation and setting of data related to files.
-   *
-   * @param objectMetadata - The metadata of the data to set.
-   * @throws ValidationException If a file identifier was not provided.
-   */
-  private ObjectStoreMetadataDto handleFileRelatedData(ObjectStoreMetadataDto objectMetadata)
-    throws ValidationException {
-    // we need to validate at least that bucket name and fileIdentifier are there
-    if (StringUtils.isBlank(objectMetadata.getBucket())
-        || StringUtils.isBlank(Objects.toString(objectMetadata.getFileIdentifier(), ""))) {
-      throw new ValidationException("fileIdentifier and bucket should be provided");
-    }
-
-    ObjectUpload objectUpload = dinaService.findOne(
-      objectMetadata.getFileIdentifier(),
-      ObjectUpload.class);
-
-    // make sure that there is an ObjectUpload that is not a derivative
-    if (objectUpload == null || objectUpload.getIsDerivative()) {
-      throw new ValidationException("primary object with fileIdentifier not found: " + objectMetadata.getFileIdentifier());
-    }
-
-    objectMetadata.setFileExtension(objectUpload.getEvaluatedFileExtension());
-    objectMetadata.setOriginalFilename(objectUpload.getOriginalFilename());
-    objectMetadata.setDcFormat(objectUpload.getEvaluatedMediaType());
-    objectMetadata.setAcHashValue(objectUpload.getSha1Hex());
-    objectMetadata.setAcHashFunction(FileController.DIGEST_ALGORITHM);
-
-    return objectMetadata;
   }
 
 }
