+++ src/main/java/ca/gc/aafc/objectstore/api/service/ObjectStoreMetaDataService.java
@@ -4,18 +4,27 @@ import ca.gc.aafc.dina.jpa.BaseDAO;
 import ca.gc.aafc.dina.service.DefaultDinaService;
 import ca.gc.aafc.objectstore.api.entities.ObjectStoreMetadata;
 import ca.gc.aafc.objectstore.api.entities.ObjectSubtype;
+import ca.gc.aafc.objectstore.api.entities.ObjectUpload;
+import ca.gc.aafc.objectstore.api.file.FileController;
 import ca.gc.aafc.objectstore.api.file.ThumbnailGenerator;
+import ca.gc.aafc.objectstore.api.respository.managedattributemap.MetadataToManagedAttributeMapRepository;
 import io.crnk.core.exception.BadRequestException;
 import lombok.NonNull;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.stereotype.Service;
 
 import javax.persistence.criteria.Predicate;
+import javax.validation.ValidationException;
+
 import java.time.OffsetDateTime;
+import java.util.Objects;
+import java.util.Optional;
 import java.util.UUID;
 
 @Service
-public class ObjectStoreMetaDataService extends DefaultDinaService<ObjectStoreMetadata> {
+public class ObjectStoreMetaDataService 
+  extends DefaultDinaService<ObjectStoreMetadata> 
+  implements ObjectStoreMetadataReadService {
 
   private final ObjectStoreMetadataDefaultValueSetterService defaultValueSetterService;
 
@@ -49,7 +58,8 @@ public class ObjectStoreMetaDataService extends DefaultDinaService<ObjectStoreMe
     if (entity.getAcSubType() != null) {
       setAcSubType(entity, entity.getAcSubType());
     }
-
+    handleFileRelatedData(entity);
+    
   }
 
   @Override
@@ -83,6 +93,10 @@ public class ObjectStoreMetaDataService extends DefaultDinaService<ObjectStoreMe
 
       setAcSubType(entity, temp);
     }
+
+    handleFileRelatedData(entity);
+    loadObjectStoreMetadata(entity.getUuid()).ifPresent(objectStoreMetadata ->
+      entity.setManagedAttribute(objectStoreMetadata.getManagedAttribute()));
   }
 
   // Overriding to implement soft delete
@@ -143,4 +157,51 @@ public class ObjectStoreMetaDataService extends DefaultDinaService<ObjectStoreMe
       .findFirst()
       .orElseThrow(() -> new IllegalArgumentException("A thumbnail subtype is not present"));
   }
+
+    /**
+   * Method responsible for dealing with validation and setting of data related to files.
+   *
+   * @param objectMetadata - The metadata of the data to set.
+   * @throws ValidationException If a file identifier was not provided.
+   */
+  private ObjectStoreMetadata handleFileRelatedData(ObjectStoreMetadata objectMetadata)
+    throws ValidationException {
+    // we need to validate at least that bucket name and fileIdentifier are there
+    if (StringUtils.isBlank(objectMetadata.getBucket())
+        || StringUtils.isBlank(Objects.toString(objectMetadata.getFileIdentifier(), ""))) {
+      throw new ValidationException("fileIdentifier and bucket should be provided");
+    }
+
+    ObjectUpload objectUpload = this.findOne(
+      objectMetadata.getFileIdentifier(),
+      ObjectUpload.class);
+
+    // make sure that there is an ObjectUpload that is not a derivative
+    if (objectUpload == null || objectUpload.getIsDerivative()) {
+      throw new ValidationException("primary object with fileIdentifier not found: " + objectMetadata.getFileIdentifier());
+    }
+
+    objectMetadata.setFileExtension(objectUpload.getEvaluatedFileExtension());
+    objectMetadata.setOriginalFilename(objectUpload.getOriginalFilename());
+    objectMetadata.setDcFormat(objectUpload.getEvaluatedMediaType());
+    objectMetadata.setAcHashValue(objectUpload.getSha1Hex());
+    objectMetadata.setAcHashFunction(FileController.DIGEST_ALGORITHM);
+
+    return objectMetadata;
+  }
+
+  @Override
+  public Optional<ObjectStoreMetadata> loadObjectStoreMetadata(UUID id) {
+    return Optional.ofNullable(this.findOne(id, ObjectStoreMetadata.class));
+  }
+
+  @Override
+  public Optional<ObjectStoreMetadata> loadObjectStoreMetadataByFileId(UUID fileId) {
+    return this.findAll(
+      ObjectStoreMetadata.class,
+      (cb, root) -> new Predicate[]{cb.equal(root.get("fileIdentifier"), fileId)}
+      , null, 0, 1)
+      .stream().findFirst();
+  }
+
 }
